Found the PDE bug--not yet fixed
Provide value of N and M as a command line argument.

Indexing out of range on multiple data structures.  This happens regardless
of whether N and M are odd, but we've only observed that errors are thrown
when compiling under g++ and N and M are odd.

I've put a bunch of assertions in place.  As an example:

Assertion `R.size()>(M*(k+1))+(j+1)' failed.